refactor(expansion-panel): add private attr.title hostbinding
